+++ .pre-commit-config.yaml
@@ -1,0 +1,17 @@
+repos:
+-   repo: git://github.com/pre-commit/pre-commit-hooks
+    sha: v1.2.0
+    hooks:
+    -   id: trailing-whitespace
+    -   id: flake8
+        args:
+        - --exclude=*/*migrations/*, dceu2019/src/dceu2019/settings/*
+    -   id: check-added-large-files
+    -   id: debug-statements
+    -   id: end-of-file-fixer
+-   repo: git://github.com/FalconSocial/pre-commit-python-sorter
+    sha: b57843b0b874df1d16eb0bef00b868792cb245c2
+    hooks:
+    -   id: python-import-sorter
+        args:
+        - --silent-overwrite
