+++ dceu2019/setup.cfg
@@ -57,7 +57,14 @@ dev =
     check-manifest==0.37
     django-debug-toolbar==1.10.1
     ipython==7.1.1
+    pre-commit
 
 [options.packages.find]
 where =
     src
+
+[flake8]
+ignore = E501
+max-line-length = 160
+max-complexity = 10
+exclude = .svn,CVS,.bzr,.hg,.git,__pycache__,.tox,.eggs,*.egg,*/*migrations,testproject
